+++ collect_app/src/main/java/org/odk/collect/android/adapters/InstanceListCursorAdapter.java
@@ -26,7 +26,7 @@ import android.widget.TextView;
 
 import org.odk.collect.android.R;
 import org.odk.collect.forms.Form;
-import org.odk.collect.android.instances.Instance;
+import org.odk.collect.forms.instances.Instance;
 import org.odk.collect.android.provider.InstanceProvider;
 import org.odk.collect.android.provider.InstanceProviderAPI.InstanceColumns;
 import org.odk.collect.android.utilities.FormsRepositoryProvider;
