+++ collect_app/src/main/java/org/odk/collect/android/dao/helpers/InstancesDaoHelper.java
@@ -17,17 +17,18 @@ package org.odk.collect.android.dao.helpers;
 import android.net.Uri;
 
 import org.odk.collect.android.application.Collect;
-import org.odk.collect.android.instances.Instance;
+import org.odk.collect.forms.instances.Instance;
 import org.odk.collect.android.javarosawrapper.FormController;
 import org.odk.collect.android.provider.InstanceProviderAPI.InstanceColumns;
 import org.odk.collect.android.utilities.InstancesRepositoryProvider;
+import org.odk.collect.forms.instances.InstancesRepository;
 
 import timber.log.Timber;
 
 /**
  * Provides abstractions over database calls for instances.
  *
- * @deprecated to favor {@link org.odk.collect.android.instances.InstancesRepository}
+ * @deprecated to favor {@link InstancesRepository}
  */
 @Deprecated
 public final class InstancesDaoHelper {
@@ -43,7 +44,7 @@ public final class InstancesDaoHelper {
      *
      * @return true if form has been marked completed, false otherwise.
      * <p>
-     * TODO: replace with method in {@link org.odk.collect.android.instances.InstancesRepository}
+     * TODO: replace with method in {@link InstancesRepository}
      * that returns an {@link Instance} object from a path.
      */
     public static boolean isInstanceComplete(boolean end, boolean completedByDefault) {
