+++ collect_app/src/main/java/org/odk/collect/android/instancemanagement/InstanceDeleter.java
@@ -2,8 +2,8 @@ package org.odk.collect.android.instancemanagement;
 
 import org.odk.collect.forms.Form;
 import org.odk.collect.forms.FormsRepository;
-import org.odk.collect.android.instances.Instance;
-import org.odk.collect.android.instances.InstancesRepository;
+import org.odk.collect.forms.instances.Instance;
+import org.odk.collect.forms.instances.InstancesRepository;
 
 import java.util.List;
