+++ collect_app/src/main/java/org/odk/collect/android/utilities/InstancesRepositoryProvider.kt
@@ -3,7 +3,7 @@ package org.odk.collect.android.utilities
 import org.odk.collect.android.application.Collect
 import org.odk.collect.android.database.DatabaseInstancesRepository
 import org.odk.collect.android.injection.DaggerUtils
-import org.odk.collect.android.instances.InstancesRepository
+import org.odk.collect.forms.instances.InstancesRepository
 import org.odk.collect.android.storage.StoragePathProvider
 
 class InstancesRepositoryProvider {
