+++ collect_app/src/test/java/org/odk/collect/android/activities/FormMapViewModelTest.java
@@ -7,8 +7,8 @@ import org.mockito.junit.MockitoJUnit;
 import org.mockito.junit.MockitoRule;
 import org.odk.collect.android.activities.viewmodels.FormMapViewModel;
 import org.odk.collect.forms.Form;
-import org.odk.collect.android.instances.Instance;
-import org.odk.collect.android.instances.InstancesRepository;
+import org.odk.collect.forms.instances.Instance;
+import org.odk.collect.forms.instances.InstancesRepository;
 import org.odk.collect.android.support.InMemInstancesRepository;
 
 import java.util.Arrays;
