+++ collect_app/src/test/java/org/odk/collect/android/provider/InstanceProviderTest.java
@@ -27,9 +27,9 @@ import static org.hamcrest.Matchers.is;
 import static org.hamcrest.Matchers.not;
 import static org.hamcrest.Matchers.notNullValue;
 import static org.hamcrest.Matchers.nullValue;
-import static org.odk.collect.android.instances.Instance.STATUS_COMPLETE;
-import static org.odk.collect.android.instances.Instance.STATUS_INCOMPLETE;
-import static org.odk.collect.android.instances.Instance.STATUS_SUBMITTED;
+import static org.odk.collect.forms.instances.Instance.STATUS_COMPLETE;
+import static org.odk.collect.forms.instances.Instance.STATUS_INCOMPLETE;
+import static org.odk.collect.forms.instances.Instance.STATUS_SUBMITTED;
 import static org.odk.collect.android.provider.InstanceProviderAPI.InstanceColumns.CONTENT_ITEM_TYPE;
 import static org.odk.collect.android.provider.InstanceProviderAPI.InstanceColumns.CONTENT_TYPE;
 import static org.odk.collect.android.provider.InstanceProviderAPI.InstanceColumns.CONTENT_URI;
